[qtmozembed] Continue functional separation between the view and the window.
Notifications that are dispatched in the context of the window should be
delivered in the context of QMozWindow, not one of the view
implementations.
