+++ src/qmozcontext.h
@@ -42,6 +42,7 @@ public:
 
 Q_SIGNALS:
     void onInitialized();
+    void destroyed();
     void recvObserve(const QString message, const QVariant data);
 
 public Q_SLOTS:
