+++ src/qmozview_defined_wrapper.h
@@ -130,8 +130,6 @@ Q_DECLARE_METATYPE(QMozReturnValue) \
     void recvMouseMove(int posX, int posY); \
     void recvMousePress(int posX, int posY); \
     void recvMouseRelease(int posX, int posY); \
-    bool Invalidate(); \
-    bool preRender();
 
 #define Q_MOZ_VIEW_SIGNALS \
     void viewInitialized(); \
