+++ src/qopenglwebpage.cpp
@@ -44,7 +44,6 @@ QOpenGLWebPage::QOpenGLWebPage(QObject *parent)
   , mCompleted(false)
   , mSizeUpdateScheduled(false)
   , mThrottlePainting(false)
-  , mReadyToPaint(true)
 {
     d->mContext = QMozContext::GetInstance();
     d->mHasContext = true;
@@ -200,21 +199,6 @@ void QOpenGLWebPage::setThrottlePainting(bool throttle)
     }
 }
 
-bool QOpenGLWebPage::readyToPaint() const
-{
-    QMutexLocker lock(&mReadyToPaintMutex);
-    return mReadyToPaint;
-}
-
-void QOpenGLWebPage::setReadyToPaint(bool ready)
-{
-    QMutexLocker lock(&mReadyToPaintMutex);
-    if (mReadyToPaint != ready) {
-        mReadyToPaint = ready;
-        Q_EMIT readyToPaintChanged();
-    }
-}
-
 /*!
     \fn void QOpenGLWebPage::initialize()
 
@@ -263,17 +247,6 @@ bool QOpenGLWebPage::event(QEvent *event)
         return QObject::event(event);
     }
     return true;
-}
-
-bool QOpenGLWebPage::Invalidate()
-{
-    return true;
-}
-
-bool QOpenGLWebPage::preRender()
-{
-    QMutexLocker lock(&mReadyToPaintMutex);
-    return mReadyToPaint;
 }
 
 bool QOpenGLWebPage::completed() const
