+++ src/qopenglwebpage.h
@@ -34,7 +34,6 @@ class QOpenGLWebPage : public QObject
     Q_PROPERTY(bool active READ active WRITE setActive NOTIFY activeChanged FINAL)
     Q_PROPERTY(bool loaded READ loaded NOTIFY loadedChanged FINAL)
     Q_PROPERTY(bool throttlePainting READ throttlePainting WRITE setThrottlePainting NOTIFY throttlePaintingChanged FINAL)
-    Q_PROPERTY(bool readyToPaint READ readyToPaint WRITE setReadyToPaint NOTIFY readyToPaintChanged FINAL)
 
     Q_MOZ_VIEW_PRORERTIES
 
@@ -63,9 +62,6 @@ public:
 
     bool throttlePainting() const;
     void setThrottlePainting(bool);
-
-    bool readyToPaint() const;
-    void setReadyToPaint(bool);
 
     void initialize();
 
@@ -97,7 +93,6 @@ Q_SIGNALS:
     void heightChanged();
     void loadedChanged();
     void throttlePaintingChanged();
-    void readyToPaintChanged();
 
     Q_MOZ_VIEW_SIGNALS
 
@@ -120,8 +115,6 @@ private:
     QMutex mGrabResultListLock;
     bool mSizeUpdateScheduled;
     bool mThrottlePainting;
-    mutable QMutex mReadyToPaintMutex;
-    bool mReadyToPaint;
 
     Q_DISABLE_COPY(QOpenGLWebPage)
 };
